+++ lib/rails_mfa.rb
@@ -1,8 +1,22 @@
 # frozen_string_literal: true
 
 require_relative "rails_mfa/version"
+require_relative "rails_mfa/configuration"
+require_relative "rails_mfa/token_manager"
+require_relative "rails_mfa/model"
+require_relative "rails_mfa/providers/base"
+require_relative "rails_mfa/providers/sms_provider"
+require_relative "rails_mfa/providers/email_provider"
 
-module RailsMfa
+module RailsMFA
   class Error < StandardError; end
-  # Your code goes here...
+
+  class << self
+    attr_accessor :configuration
+
+    def configure
+      self.configuration ||= Configuration.new
+      yield(configuration) if block_given?
+    end
+  end
 end